Ensure the declared type of a formal is, indeed, a type
This extends the checking introduced in #20262 to all formals.

After making a big deal out of it here:
  https://github.com/chapel-lang/chapel/pull/20262#issuecomment-1190259631

I realized that maybe it is actually simple? Maybe indeed all we need
is to add a field `bool typeExprFromDefaultExpr` to `ArgSymbol`?

Turned out that it was not so simple after all. On the upside, I was
able to remove some heuristic code, replacing it with a
(non-heuristic) check for `typeExprFromDefaultExpr`.

It took some work to handle the various ways a user may declare a
formal's type, see `checkFormalType()`. I think that this code will
still miss some cases ex. when the formal's type is declared using a
value-producing primitive. The way it should be done is by flagging
the cases where the `qualType()` of the formal's typeExpr is not a
QUAL_TYPE. However, our qualType() does not have the QUAL_TYPE option,
so we cannot do this at present. Luckily, Dyno gets it right with
`QualifiedType::Kind::TYPE` .

While there, I added two futures asking about the behavior in some cases
where the formal's type is a runtime type, specifically an
array. These futures support the cause of #19292, showing another
example of where runtime types add complexity, albeit minor in this
case, to both the implementation and the language definition.

Signed-off-by: Vassily Litvinov <vasslitvinov@users.noreply.github.com>