Allow gpu kernels to call the CHPL_RT_ERROR (halt) primitive
Add an integer argument to the end of the kernel's parameter list that can be
set to a non-zero value if a halt primitive is encountered. After outlining
the kernel, find any calls to halt and replace them with moves that set
the new argument to non-zero.

In the gpu runtime allocate one extra space in the kernel_params array
for the new argument and set it to point at a gpu-allocated int. When the
kernel completes, read the value and issue an error if it was non-zero.

Add a simple test that can either call halt or not based on a config const.

Signed-off-by: David Iten <daviditen@users.noreply.github.com>