Rework mason documentation to be multi-page This PR changes the documentation for mason located at tools/mason to be multi-document documentation, rather than just one huge wall of text.
Also in this PR is a move from having two distinct
mason documentations, one located at tools/mason and the
other located at writing-chapel-programs/mason-packages, to
be located in one central location, which is
writing-chapel-programs/mason-packages. This choice was made
because that documentation is much easier to locate than
the tools/mason link. The documentation at tools/mason now
just points to the other mason documentation.

Signed-off-by: Ben McDonald <mcdobe100@gmail.com>