Add implementations of some methods based on experiment with locale hashing
Takes lessons learned from my side program experiment and uses them to
implement:

- the initializers and fields needed
- contains
- add
- set

Also makes the function used for default hashing be a method, since it can't be
generic and stored in the field normally.

Note that this current implementation for these functions have some caveats but
that I think they're a reasonable implementation start.

Caveats:
- Duplicates the logic to determine the locale for the key.  Since we could be
using our default method or a user provided one, this is 6 lines per function
and so probably should be put into a helper method
- Every one of these methods locks inside its body on the locale.  I think it
is useful to provide the individual version like this, though it is slow.  But
for aggregation to occur, we'll need a hidden version that does not lock itself
that the aggregator can call after triggering the locking itself.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>