Update to TupleType to support VarArgs
Introduces two new fields to TupleType: isVarArgTuple_ and isKnownSize_.

The 'isVarArgTuple_' field may be useful for const-checking in the
future, when determining the appropriate kind of an expression like
`args(i)`.

The 'isKnownSize_' bool allows us to represent varargs with a type
expression but not a count expression. Such a tuple can be created with
a substitution with the postOrderId of '-1' and a QualifiedType
representing the star-type.

Signed-off-by: Ben Harshbarger <ben.harshb@gmail.com>