Update resolveNamedDecl to handle VarArgFormals
There are two important parts of this change.

First, when resolving with an instantiated type we need to update the
intents for the tuple elements because they are sometimes based on the
actual's QualifiedType rather than the formal's. This step is handled by
the 'rebuildVarArgSub' helper function.

Second is the non-instantiating case where we need to build a
VarArgTuple from the computed kind and type-expression, and is handled
by the 'computeVarArgTuple' helper function. The type-expression used is
what the user specified *per-element*. For example, in "args:int...",
this helper function will have the type 'int' and need to build a
star-tuple of unknown size. The resulting type of this function can be
erroneous if the count-expression is invalid ('isValidVarArgCount').

Signed-off-by: Ben Harshbarger <ben.harshb@gmail.com>