Update FormalActualMap::computeAlignment to handle VarArgs
Two fields are added to the FormalActual entry type, one to track
whether a formal has a default, and another to track whether an entry is
part of a VarArgFormal.

The approach taken here is to create a FormalActual entry for each
actual that the caller attempted to pass to the VarArgFormal. Currently
this method is usually called before we have evaluated the types of
formals, so we cannot measure alignment based on a possible
count-expression. Each entry for the VarArgFormal will keep track of the
VarArg's type expression and its own actual type, and will point to the
same formal in the uAST.

Signed-off-by: Ben Harshbarger <ben.harshb@gmail.com>