Update instantiateSignature to handle VarArgFormals
The FormalActualMap treats each argument to the VarArgFormal as its own
entry, so to actually instantiate the TupleType we want we need to store
each computed type and whether that type required instantiation. If any
argument to the VarArgFormal requires instantiation, or if the size of
the VarArg is not known, then we should instantiate with a TupleType.

Signed-off-by: Ben Harshbarger <ben.harshb@gmail.com>