Add tests for resolving VarArgs
This testing framework for VarArgFormals currently includes 41 test
cases, with optional output determined by the arguments --debug and
--verbose. This extra output is disabled by default.

The primary helper function in this framework is 'testMatcher', which
accepts an intent, type, and count for the VarArgFormal, and a struct
representing the arguments to the VarArgs. This helper function will
build a string that represents a valid Chapel program. This program will
contain two functions, one with a VarArgFormal and another with
individual formals that share the intent and type provided for the
VarArgFormal.

Once the program is generated, 'testMatcher' will parse and resolve the
program. It will then pass the resolved results into a ResolvedVisitor
that wraps around a Collector type built for this test. The Collector
type will gather names for declarations (i.e. formals) and identifiers.
This this information, 'testMatcher' will call the 'validate' helper
function which will compare the resolved types between the VarArg
function and the 'normal' function. 'validate' will also check type
queries and the count-expression.

Another helper function is 'customHelper', which is meant to accept an
already-built program, resolve it, run the Collector, check for errors,
and return the Collector results. This version is more useful for some
cases that don't make sense to compare against a 'normal' function, like
checking for specific values in the count expression, or errors for
multiple VarArgFormals.

The initial test cases are:
- passing an int, real, and string to fully-generic "args..."
  - passed to various intents (default,in,ref,const-ref,const,const-in)
  - passed as param values or types to param/type intent

- passing int(8), int, and int(32) to "args:int(?)..."

- passing different types to a type query "args:?t..."
  - also includes params

- passing various concrete types to concrete type expressions
  - ints to "args:int..."
  - int, int(8), and int(32) to "args:int..."
  - 3*int tuples to "args:(int,int,int)..."

- various count-expressions
  - '3', '?', '?n'
  - "args:?t...?n", "args:?t...3"

- use a param formal to constrain the number of VarArgs
  - 'param n : int, args...n'

- various alignment tests
  - named arguments
  - typed formals after the VarArgs

These are the cases that are supposed to encounter errors:
- passing string to "args:int(?)..."
- pass vars to param/type intent
- pass non-homogenous types to a type-query "args:?t..."
  - also "args:int(?n)..."
- pass int to "args:int(8)..."
- pass 3*int to "args:(int(8),int,int)..."
- passing wrong number of arguments to count-constrainted VarArgs
  - when constrained literals or params
- specify zero VarArgs: "args...0"
- passing zero arguments to "args..."

Signed-off-by: Ben Harshbarger <ben.harshb@gmail.com>