Implement and add a test for the `clear` method.  Make distributedMap a record
When testing `clear`, I discovered that it was problematic for distributedMap
to be a class with that method, as `owned` and `shared` both have their own
version and we resolve to that one first rather than forwarding to the one
defined on the particular class type.  I opened an issue to discuss this but in
the meanwhile, change distributedMap to also be a record so that the test
exercises what I was intending.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>