Resolve expression-level conditions
When treated like a generic `AstNode`, the `Condition` node
does not get assigned a `QualifiedType` (much like `Block` and similar).
However, `Condition` _should_ have a type when used as an expression,
such as in the following case (adapted from `ChapelRange.range`).

    var x : if someCondition then oneType else anotherType;

In this case, we want `x` to be of type given by either `oneType`
or `anotherType`, and not `Unknown` (what it is at present). To this
end, it's necessary to unwrap the statements within both the
`then` and `else` blocks, and propagate their type up to the whole
condition. This PR also adds short-circuiting for conditions.

Signed-off-by: Danila Fedorin <daniel.fedorin@hpe.com>