Fixed signed/unsigned comparison when computing number of bytes to read.
When computing the number of bytes remaining to be read in the file the
current file offset was subtracted from the current file length. The
length may be less than the offset because the file size may change.
Previously, the difference was assigned to an unsigned integer that was
compared against zero, leading to a conditional that was always false.
Now the comparison is done prior to the assignment.

Signed-off-by: John H. Hartman <jhh67@users.noreply.github.com>