Add checking of TOML during `mason add` Previously, the `mason add` command would add anything that was passed to it, regardless of whether or not the package existed.
This PR modifies `mason add` to check the TOML file of the
package that was `mason add`ed prior to adding it to the manifest
file and also adds a check to ensure that the type of the project
that is being added was a mason library.

As of recently, the manifest file has a new `type` field, and
projects with `type="library"` are the only type that can be added
as a dependency of a project.

Signed-off-by: Ben McDonald <mcdobe100@gmail.com>