Infer return types from multiple returns
This commit adds code that's used to reconcile multiple return
types. This happens by trying to find the most general kind / intent
for a set of types, then trying to find one type that the rest
can all be converted to. This is done using the usual formal/actual
passing rules.

Signed-off-by: Danila Fedorin <daniel.fedorin@hpe.com>