Initialize a private array constant with a tuple instead of array
Now that DateTime is combined into Time a test that does 'use Time'
and tracking bulk transfers is seeing an extra array transfer. This is from
initializing a private const array from DateTime with an array literal.
Replace the literal with a tuple literal instead to avoid the extra transfer
message.

Signed-off-by: David Iten <daviditen@users.noreply.github.com>