Clean up tests to use commonType directly.
This was a very good idea, thanks @mppf. Besides making the tests
cleaner, it also helps decouple commonType from its use in function
return type computations. This, in turn, makes it possible to test
cases where a return intent is not enforced upon the list of types,
which would be the case in if-expressions (say).

Signed-off-by: Danila Fedorin <daniel.fedorin@hpe.com>