Add prototype handling for AMD GPU target.
This mirrors the CUDA compilation process; much like
CUDA, we fork, use LLVM to generate code for our device,
and then combine the device code and host code. Unlike CUDA,
we can generate an object file right away instead of generating
an assembly file and running it through an assembler.

Signed-off-by: Andy Stone <stonea@users.noreply.github.com>
Signed-off-by: Danila Fedorin <daniel.fedorin@hpe.com>