Finish removing test_many_ranks_of_arithmetic_domains.chpl
In PR #20440, I believe Andy intended to remove this test based on the
OP and our discussion, but my review was sloppy and I didn't notice it
hadn't been.  This caused it to fail.  Removing it here, as I believe
was intended to resolve last night's failures.  If I'm incorrect/have
misunderstood, we can fix it in a different way on Monday.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>