Optimize distributed rectangular domain creation
When creating distributed rectangular domains, directly create the final
distributed domain instead of creating an empty domain and assigning the
local domain into it. This eliminates a distributed domain assignment
and re-privatization, which gets rid of 3 coforall+ons resulting in some
domain creation speedups.

Here's the performance results for creating a block distributed domain
on 16 nodes of a CS and XC:

```
./distCreate --size=arraySize.tiny --dist=distType.block -nl 16
```

16-node-cs-hdr

| config | domInit  |
| ------ | -------: |
| before | 0.00464s |
| after  | 0.00288s |

16-node-xc:

| config | domInit  |
| ------ | -------: |
| before | 0.00055s |
| after  | 0.00027s |

Part of 20197

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>