Update file.reader to take a region argument (and avoid a resolution conflict)
Removes the default value for the hint argument on the overload with the iohints
deprecation warning.  Also extends the deprecation test to handle the additional
deprecation warnings.

This also involved compacting openreader, now that it is calling something that
takes a region argument instead of a start and end argument.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>