Feed warning flags to invocations of C++ compiler
Prior to this commit, C++ files in Dyno did not receive WARN_CXXFLAGS,
and thus did not always throw warnings. This made it possible to push
code past CI that did not work on some machines due to -Werror.
This commit ensures that warning flags are fed to the C++ compiler.

Signed-off-by: Danila Fedorin <daniel.fedorin@hpe.com>