Update MatrixMarket and URL packages for file.reader interface changes
I might want to update the URL functions `openUrlReader` and `openUrlWriter` to
use a region argument themselves, but that might be a distraction?  I did update
them about the style argument but that's because the type it took got
deprecated.  This module isn't up for 2.0 stabilization, but it might be nice
to keep it consistent.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>