Scope-resolve task-intents in dyno
Updates scope-queries to step into WithClauses to find task-intents,
which are represented as NamedDecls.

The Resolver is updated to store the aliased ID in ResolvedExpression
for later use. This requires performing a lookup in the scope outside of
the loop-statement.

Also sets the type for simple task-intents that use an IntentList.
Resolving the type for reduce-intents is not supported at this time.

Signed-off-by: Ben Harshbarger <ben.harshb@gmail.com>