Adding Tests for the Unstable Warning
test1.chpl test1.good:The unstable keyword is trigged for a config variable.
tetst3.chpl test3.good:The unstable keyword is triggered whenever the record 'Test3' is called.
test4.chpl test4.good:The unstable keyword is placed for a statement, rather than symbol. Therefore, it will be a syntax error.

Signed-off-by: Nikki Rad <nikki.rad@hpe.com>