Hide some compile options from clang-based linter
The compile_commands.json fed to the linter are generated
by whatever compiler we used to build the host. However,
the linter is always run using Clang. This means that some
flags that exist in GCC but not in Clang are fed to the
linter, which leads to warnings. This adjusts the linting
script to strip these flags from compile_commands.

Signed-off-by: Danila Fedorin <daniel.fedorin@hpe.com>