Add allNumericsBinaryRealImagComplex-errors test
This test results in an ambiguity error for f(myInt8, myInt64).

In principle, we could adjust the disambiguation rules to prefer the
real(64) version, by adding another stage to comparing arguments, but
that would make the disambiguation rules more complex & I don't think we
have a particularly good reason that this case shouldn't be ambiguity.

---
Signed-off-by: Michael Ferguson <mppf@users.noreply.github.com>