Update the names of errors with acronyms from 'SysError'
There were 4 errors in the 'SysError' module (now part of 'OS') that had
capitalized acronyms as part of their names. Change them to use regular
camelCase for the acronyms.

Add deprecated versions of the original names, update tests and modules
to use the new names, add a test/deprecated test that gets deprecated warnings
for using the old names.

Signed-off-by: David Iten <daviditen@users.noreply.github.com>