Make params without values be considered unknown
This is needed for the resolution of if-expressions with `param`
conditions. An if-expression depending on a `param bool` without a value
should be considered unknown (as is the case for an if-expression
depending on any other unknown value).

This commit also implements special-case behavior for `param _tuple`
types, which have their tuple elements' values encoded in the
`QualifiedType`s of the `TupleType`, but not stored into a `Param*`.

Signed-off-by: Danila Fedorin <daniel.fedorin@hpe.com>