Adjust range intersection to not overload min/max
This code would create overload sets errors, presumably,
if overloads sets errors were enabled for internal modules.

Anyway, I am adjusting when visibility is considered in disambiguation
and I don't know of a good reason that this code needs to define inner
functions named 'min' and 'max', so this commit renames these
and adds a forwarding overload to the regular min/max.

---
Signed-off-by: Michael Ferguson <mppf@users.noreply.github.com>