Adjust the Python 3.5 job to behave more like the Python 2.7 job
Adds a flag to nightly that can be used with jobs to test versions of Python 3
that are no longer supported by chpldoc and start_test.  This is separate from
the python2 flag, since Python 2 can have additional constraints.  Use this flag
with the job tracking Python 3.5.

This should avoid the sporadic failures we've been seeing when the job happens
to run with the version of Python it is supposed to.  Other changes will be
necessary to avoid running places where Python 3.5 doesn't exist and to ensure
that the correct version of Python is used, but I think this should get in
independently of those.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>