Don't count the comms from end counts
Don't track communication coming from end counts, which are used to
track task completion. End counts use an `atomic int` to count tasks, and
atomics are implemented differently depending on `CHPL_NETWORK_ATOMICS`.
This leads to different comm count output (tracking fast-ons/ons vs.
amos), which increases the cost of maintaining comm count tests. We also
already track the comm that's used to initiate the remote tasks
separately (as non-blocking ons) so this change doesn't hide that
communication is happening, we're just papering over some of the
implementation details. And note that we already do this blocking-ons,
it's just that the completion comms for that are entirely within the
comm layer whereas end counts tracking for remote tasks is done in
module code so we have to disable the tracking manually. To do that,
this uses new functionality to disable comm diags for a specific task,
wrapped in a context manager to make the callsites cleaner

Here's the comm diags before and after for a no-op coforall+on:

```chpl
coforall loc in Locales do on loc {}
```

Processor atomics before:

| locale | execute_on_fast | execute_on_nb |
| -----: | --------------: | ------------: |
|      0 |               0 |             2 |
|      1 |               1 |             0 |
|      2 |               1 |             0 |

Network atomics before:

| locale |      amo | execute_on_nb |
| -----: | -------: | ------------: |
|      0 | unstable |             2 |
|      1 | unstable |             0 |
|      2 | unstable |             0 |

Network atomics before (`-scommDiagsPrintUnstable=true`):

| locale | amo | execute_on_nb |
| -----: | --: | ------------: |
|      0 |  61 |             2 |
|      1 |   1 |             0 |
|      2 |   1 |             0 |

All configs now:

| locale | execute_on_nb |
| -----: | ------------: |
|      0 |             2 |
|      1 |             0 |
|      2 |             0 |

This allows us to get rid of almost all our network-atomic comm count
good files since the only difference in them was comm coming from end
counts.

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>