Update tests to use OS.errorCode instead of SysBasic.syserr
Update tests to use OS.errorCode instead of SysBasic.syserr.
The only non-obvious change in these tests is for
modules/sungeun/init/printModuleInitOrder. There the OS and POSIX
modules were removed from the expected output. This is because an
internal module now uses them first and so they're already initialized
before the code to print the order runs.

Signed-off-by: David Iten <daviditen@users.noreply.github.com>