Fix problems with two mandelbrot tests
The generic function e.g. 'operator +(cr, ci)' was being
preferred because it is closer in terms of visibility
and is a candidate for 'myReal + myReal' etc.

To fix, add a bit more types to it, so it is not a candidate
for the regular '+' it is trying to replace.

---
Signed-off-by: Michael Ferguson <mppf@users.noreply.github.com>