Support mixed array literals as long as they're assignable
Prior to this PR, an array literal like '[1.0, 2]' would not work
because the compiler would infer the type of the array based on
the initial element (real) and then say that the type of '2' (int)
didn't match.

This commit loosens that logic in the event that the subsequent
elements can be assigned to a variable of the first one's type.
Thus '[1.0, 2]' would be inferred to be an array of reals, but
'[1, 2.0]' would still cause problems since it would be
inferred to be an array of ints, and we  can't assign reals to
ints.

There's also some commented out code here that uses some clever
meta-programming to try and determine the common type (such that
either would be inferred to be an array of reals), but it seems likely
to blow up compile times and also of generating terrible error
messages if the types can't be unified.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>