Update ranges/domains w.r.t. things I've deprecated
* removed the deprecated sizeReturnsInt config param
* made range low / high bounds return aligned bounds by default
  - deprecated 'alignedBoundsByDefault' config param since it's no longer used
  - updated documentation to reflect new behavior

Note that this keeps the alignedLow/alignedBounds queries on ranges
for now.  I'd imagined we would deprecate these now that they're
duplicates, but at least one user was potentially interested in
keeping them in order to have a more descriptive way of saying what
you wanted.  I'm not sure I agree with this, but it deserves more
discussion than we've given it.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>