Update spectral norm perfkeys with correct line number
PR #20547 eliminated the deprecation warning in these tests, but
did not account for the change in line numbers which affected the
verify step of -performance testing. This PR updates the perfkeys
to look at the first line rather than the fourth, now that the
first three lines no longer consist of a deprecation warning.

Signed-off-by: Ben Harshbarger <ben.harshb@gmail.com>