Add overlappedArrayAssignment.chpl test
For CHPL_GPU_MEM_STRATEGY=array_on_device.

I'm kind of suprised this doesn't fail since I wouldn't expect cuMemcpyDtoD to
be careful about not overwriting memory when array regions overlap.

But hey, it works. Let's lock down this behavior in case it regresses or we
want to test it with different GPU runtimes (like for AMD in the future).

---
Signed-off-by: Andy Stone <stonea@users.noreply.github.com>