Adjust function instantiation to enforce type queries
Currently, type query constraints are not enforced when a function
signature is instantiated. In particular, the following example works,
when it should not.

```Chapel
record R { type a; type b; }
proc f(r: R(?t, t)) {}
var x: R(int, string);
f(x); // shouldn't work, int != string
```

The problem is that type query constraints are not enforced, because
compatibility of formal and actual types is only computed once, and the
resulting instantiations are not verified.

This commits adjusts the behavior by modifying `instantiateSignature` in
the following ways.
1. It moves the second ->traverse() pass of each formal into
   the first traversal loop. The second ->traverse() pass is needed to
   re-compute intents (and possibly other things, see `resolveNamedDecl`).
   This second pass is also used to compute type query information.
2. The type query information is used to re-compute the formal's type
   again. To accommodate this, the resolver was modified with two
   additional flags, one for ignoring substitutions, and one for
   skipping re-computing type query information. With these flags, a
   "proper" formal type is computed, which includes information from the
   updated type queries, while not being poisoned by existing
   substitutions.

   This re-computed type is then checked for compatibility with the
   instantiated formal type. This ensures the latter type satisfies the
   type query constraints.
3. A new resolver is no longer created for re-computing intents, as this
   information is now determined via step (1).

Signed-off-by: Danila Fedorin <daniel.fedorin@hpe.com>