Replace `.skipif` with `.suppressif`
The affected tests used to be skipped when testing the `--baseline`
configuration because they contained throwing iterators.
Which is currently not implemented without inlining and so
produces a compilation error upon `--baseline`.

Replace their skipifs with identical suppressifs, so that
if/when this feature gets implemented, we will be notified. See
https://github.com/chapel-lang/chapel/pull/20564#issuecomment-1233226030

While there, make all suppressifs identical.
Alas this throws off git diff.

Suggested by @bradcray, not reviewed.

Signed-off-by: Vassily Litvinov <vasslitvinov@users.noreply.github.com>